language-app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
boto3
gtts